refactor(nlu): moved intent election + ambiguity in own middleware
