+++ modules/.knowledge/src/backend/middleware.ts
@@ -29,7 +29,7 @@ export const registerMiddleware = async (bp: typeof sdk, classifiers: Classifier
 
       next()
     },
-    order: 15,
+    order: 16,
     description: 'Finds content from Knowledge base files'
   })
 }
