+++ modules/analytics/src/backend/setup.ts
@@ -60,7 +60,7 @@ export default async (bp: typeof sdk, db: Database, interactionsToTrack: string[
     name: 'analytics.incoming',
     direction: 'incoming',
     handler: incomingMiddleware,
-    order: 12, // after nlu and qna
+    order: 14, // after nlu and qna
     description: 'Tracks incoming messages for Analytics purposes'
   })
