+++ modules/nlu/src/backend/predict-pipeline.ts
@@ -7,7 +7,6 @@ import LanguageIdentifierProvider, { NA_LANG } from './language/language-identif
 import { isPOSAvailable } from './language/pos-tagger'
 import { getUtteranceFeatures } from './out-of-scope-featurizer'
 import SlotTagger from './slots/slot-tagger'
-import * as math from './tools/math'
 import { replaceConsecutiveSpaces } from './tools/strings'
 import { EXACT_MATCH_STR_OPTIONS, ExactMatchIndex, TrainArtefacts } from './training-pipeline'
 import { Intent, PatternEntity, SlotExtractionResult, Tools } from './typings'
@@ -61,8 +60,6 @@ type E1IntentPred = {
 
 const DEFAULT_CTX = 'global'
 const NONE_INTENT = 'none'
-const OOS_AS_NONE_TRESH = 0.4
-const LOW_INTENT_CONFIDENCE_TRESH = 0.4
 
 async function DetectLanguage(
   input: PredictInput,
@@ -290,127 +287,6 @@ async function predictIntent(input: PredictStep, predictors: Predictors): Promis
   }
 }
 
-// taken from svm classifier #295
-// this means that the 3 best predictions are really close, do not change magic numbers
-function predictionsReallyConfused(predictions: sdk.MLToolkit.SVM.Prediction[]): boolean {
-  if (predictions.length <= 2) {
-    return false
-  }
-
-  const std = math.std(predictions.map(p => p.confidence))
-  const diff = (predictions[0].confidence - predictions[1].confidence) / std
-  if (diff >= 2.5) {
-    return false
-  }
-
-  const bestOf3STD = math.std(predictions.slice(0, 3).map(p => p.confidence))
-  return bestOf3STD <= 0.03
-}
-
-// TODO implement this algorithm properly / improve it
-// currently taken as is from svm classifier (engine 1) and doesn't make much sens
-function electIntent(input: PredictOutput): PredictOutput {
-  const allCtx = Object.keys(input.predictions)
-
-  const ctx_predictions = allCtx.map(label => {
-    const { confidence } = input.predictions[label]
-    return { label, confidence }
-  })
-
-  const perCtxIntentPrediction = _.mapValues(input.predictions, p => p.intents)
-
-  const oos_predictions = _.mapValues(input.predictions, p => p.oos)
-
-  const totalConfidence = Math.min(
-    1,
-    _.sumBy(
-      ctx_predictions.filter(x => input.includedContexts.includes(x.label)),
-      'confidence'
-    )
-  )
-  const ctxPreds = ctx_predictions.map(x => ({ ...x, confidence: x.confidence / totalConfidence }))
-
-  // taken from svm classifier #349
-  let predictions = _.chain(ctxPreds)
-    .flatMap(({ label: ctx, confidence: ctxConf }) => {
-      const intentPreds = _.chain(perCtxIntentPrediction[ctx] || [])
-        .thru(preds => {
-          if (oos_predictions[ctx] >= OOS_AS_NONE_TRESH) {
-            return [
-              ...preds,
-              {
-                label: NONE_INTENT,
-                confidence: oos_predictions[ctx],
-                context: ctx,
-                l0Confidence: ctxConf
-              }
-            ]
-          } else {
-            return preds
-          }
-        })
-        .map(p => ({ ...p, confidence: _.round(p.confidence, 2) }))
-        .orderBy('confidence', 'desc')
-        .value() as (sdk.MLToolkit.SVM.Prediction & { context: string })[]
-      if (intentPreds[0]?.confidence === 1 || intentPreds.length === 1) {
-        return [{ label: intentPreds[0].label, l0Confidence: ctxConf, context: ctx, confidence: 1 }]
-      }
-
-      const noneIntent = { label: NONE_INTENT, context: ctx, confidence: 1 }
-      if (predictionsReallyConfused(intentPreds)) {
-        intentPreds.unshift(noneIntent)
-      }
-
-      if (intentPreds.length <= 1) {
-        return noneIntent
-      }
-
-      const lnstd = math.std(intentPreds.filter(x => x.confidence !== 0).map(x => Math.log(x.confidence))) // because we want a lognormal distribution
-      let p1Conf = math.GetZPercent((Math.log(intentPreds[0].confidence) - Math.log(intentPreds[1].confidence)) / lnstd)
-      if (isNaN(p1Conf)) {
-        p1Conf = 0.5
-      }
-
-      return [
-        { label: intentPreds[0].label, l0Confidence: ctxConf, context: ctx, confidence: _.round(ctxConf * p1Conf, 3) },
-        {
-          label: intentPreds[1].label,
-          l0Confidence: ctxConf,
-          context: ctx,
-          confidence: _.round(ctxConf * (1 - p1Conf), 3)
-        }
-      ]
-    })
-    .orderBy('confidence', 'desc')
-    .filter(p => input.includedContexts.includes(p.context))
-    .uniqBy(p => p.label)
-    .map(p => ({ name: p.label, context: p.context, confidence: p.confidence }))
-    .value()
-
-  const ctx = _.get(predictions, '0.context', 'global')
-  const shouldConsiderOOS =
-    predictions.length &&
-    predictions[0].name !== NONE_INTENT &&
-    predictions[0].confidence < LOW_INTENT_CONFIDENCE_TRESH &&
-    oos_predictions[ctx] > OOS_AS_NONE_TRESH
-  if (!predictions.length || shouldConsiderOOS) {
-    predictions = _.orderBy(
-      [
-        ...predictions.filter(p => p.name !== NONE_INTENT),
-        { name: NONE_INTENT, context: ctx, confidence: oos_predictions[ctx] || 1 }
-      ],
-      'confidence'
-    )
-  }
-
-  const elected = _.maxBy(predictions, 'confidence')
-  return {
-    ...input,
-    intent: elected,
-    intents: predictions
-  }
-}
-
 async function predictOutOfScope(input: PredictStep, predictors: Predictors): Promise<PredictStep> {
   const oos_predictions = {}
   if (!isPOSAvailable(input.languageCode) || _.isEmpty(predictors.oos_classifier_per_ctx)) {
@@ -441,22 +317,6 @@ async function predictOutOfScope(input: PredictStep, predictors: Predictors): Pr
   }
 }
 
-function detectAmbiguity(input: PredictOutput): PredictOutput {
-  // +- 10% away from perfect median leads to ambiguity
-  const preds = input.intents
-  const perfectConfusion = 1 / preds.length
-  const low = perfectConfusion - 0.1
-  const up = perfectConfusion + 0.1
-  const confidenceVec = preds.map(p => p.confidence)
-
-  const ambiguous =
-    preds.length > 1 &&
-    (math.allInRange(confidenceVec, low, up) ||
-      (preds[0].name === NONE_INTENT && math.allInRange(confidenceVec.slice(1), low, up)))
-
-  return { ...input, ambiguous }
-}
-
 async function extractSlots(input: PredictStep, predictors: Predictors): Promise<PredictStep> {
   const slots_per_intent: typeof input.slot_predictions_per_intent = {}
   for (const intent of predictors.intents.filter(x => x.slot_definitions.length > 0)) {
@@ -465,16 +325,6 @@ async function extractSlots(input: PredictStep, predictors: Predictors): Promise
   }
 
   return { ...input, slot_predictions_per_intent: slots_per_intent }
-}
-
-function extractElectedIntentSlot(input: PredictOutput): PredictOutput {
-  const intentWasElectedWithoutAmbiguity = input?.intent?.name && !_.isEmpty(input.predictions) && !input.ambiguous
-  if (!intentWasElectedWithoutAmbiguity) {
-    return input
-  }
-
-  const electedIntent = input.predictions[input.intent.context].intents.find(i => i.label === input.intent.name)
-  return { ...input, slots: electedIntent.slots }
 }
 
 function MapStepToOutput(step: PredictStep, startTime: number): PredictOutput {
@@ -497,20 +347,6 @@ function MapStepToOutput(step: PredictStep, startTime: number): PredictOutput {
         }
       } as sdk.NLU.Entity)
   )
-  // legacy pre-ndu
-  const slots = step.utterance.slots.reduce((slots, s) => {
-    return {
-      ...slots,
-      [s.name]: {
-        start: s.startPos,
-        end: s.endPos,
-        confidence: s.confidence,
-        name: s.name,
-        source: s.source,
-        value: s.value
-      } as sdk.NLU.Slot
-    }
-  }, {} as sdk.NLU.SlotCollection)
 
   const slotsCollectionReducer = (slots: sdk.NLU.SlotCollection, s: SlotExtractionResult): sdk.NLU.SlotCollection => {
     if (slots[s.slot.name] && slots[s.slot.name].confidence > s.slot.confidence) {
@@ -604,11 +440,7 @@ export const Predict = async (
     stepOutput = await predictIntent(stepOutput, predictors)
     stepOutput = await extractSlots(stepOutput, predictors)
 
-    let predictOutput = MapStepToOutput(stepOutput, t0)
-    predictOutput = electIntent(predictOutput)
-    predictOutput = detectAmbiguity(predictOutput)
-    predictOutput = extractElectedIntentSlot(predictOutput)
-    return predictOutput
+    return MapStepToOutput(stepOutput, t0)
   } catch (err) {
     if (err instanceof InvalidLanguagePredictorError) {
       throw err
