+++ modules/qna/src/backend/setup.ts
@@ -25,7 +25,7 @@ export const initModule = async (bp: typeof sdk, bots: ScopedBots) => {
         next()
       }
     },
-    order: 11, // must be after the NLU middleware and before the dialog middleware
+    order: 13, // must be after the NLU middleware and before the dialog middleware
     description: 'Listen for predefined questions and send canned responses.'
   })
